app/requirements.txt
pyTelegramBotAPI
airtable-python-wrapper
requests